Adding files
There's a lot to be decided on. I'm adding in a compiled version of FFTW
along with the source for compiling it using the Superpowers/build-libs.sh
script (see the source for the environment vars it expects). I've added
the kissFFT source, which needs an Android.mk.

You'll need to run `andproid update project` to be able to compile this on
your compy (see the docs for the right commond line options).

Then you should be able to run `ant debug` to build a debug copy.

To put it on your device run `adb install bin/Superpowers-debug.apk`.